New way of history logging.
History objects are now stored on a Player. It's separate from the 'Player' table to reduce the lookup time for rendering the Game-Table (other wise this lookup would get all a Players history every time the app loads.)

All the other History that has been logged since then will be lost.

When doing Stats Graphing you will then be able to do a quicker lookup for a Players entire history.